Added rest of url to API_PATH
This fixes what appeared to be a bug where this library was passing an
incomplete url to the Oauth library.

The test suite does not adequatly convey whether there is some other desired
design meant to be at work here.  I can imagine a world where the oauth
library automatically provides the first part (scheme+host) of the url,
however, that wasn't what I was observing.